+++ example/lib/main.dart
@@ -1,15 +1,6 @@
 import 'dart:developer';
 
-import 'package:df_dropdown/df_searchable_dropdown.dart';
-import 'package:df_dropdown/df_searchable_multi_select_dropdown.dart';
-import 'package:df_dropdown/df_searchable_single_select_dropdown.dart';
-import 'package:df_dropdown/df_simple_dropdown.dart';
-import 'package:df_dropdown/enums/dropdown_type.dart';
-import 'package:df_dropdown/models/drop_down_model.dart';
-import 'package:df_dropdown/models/dropdown_decoration.dart';
-import 'package:df_dropdown/models/multi_selector_decoration.dart';
-import 'package:df_dropdown/models/simple_selector_decoration.dart';
-import 'package:df_dropdown/models/single_selector_decoration.dart';
+import 'package:df_dropdown/df_dropdown.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -39,12 +30,56 @@ class MyApp extends StatelessWidget {
                 clipBehavior: Clip.none,
                 child: Column(
                   mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
+                    DfDropdownWrapper(
+                      child: Row(
+                        children: [
+                          SizedBox(
+                            width: 4,
+                          ),
+                          Icon(Icons.car_rental),
+                          SizedBox(
+                            width: 4,
+                          ),
+                          Text("Custom Dropdown Example"),
+                        ],
+                      ),
+                      decoration: DropdownDecoration(
+                        backgroundColor: Colors.greenAccent,
+                      ),
+                      selectorDecoration: SimpleSelectorDecoration(
+                        selectedItemIcon: Icon(
+                          Icons.cabin,
+                        ),
+                        selectedItemColor: Colors.blue.withOpacity(
+                          0.4,
+                        ),
+                      ),
+                      onOptionSelected: (option) {
+                        log("Option selected ${option}");
+                      },
+                      selectedValue: DropDownModel<String>(
+                        key: "1",
+                        value: "1",
+                        text: "New York City",
+                      ),
+                      initData: [
+                        DropDownModel<String>(
+                            key: "1", value: "1", text: "New York City"),
+                        DropDownModel<String>(
+                            key: "2", value: "2", text: "Los Angeles"),
+                      ],
+                    ),
+                    const SizedBox(
+                      height: 8,
+                    ),
                     DfSimpleDropdown<String>(
                       dropdownType: DropdownType.overlay,
                       decoration: DropdownDecoration(
                         borderRadius: BorderRadius.circular(999),
                         borderColor: Colors.blue,
+                        backgroundColor: Colors.green.withOpacity(0.2),
                       ),
                       selectorDecoration: SimpleSelectorDecoration(
                         selectorColor: Colors.amber.shade300,