+++ lib/df_dropdown_wrapper.dart
@@ -1,21 +1,19 @@
-import 'package:df_dropdown/enums/dropdown_type.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '/models/drop_down_model.dart';
 import '/models/dropdown_decoration.dart';
 import '/models/simple_selector_decoration.dart';
-import '/widgets/dropdown_field.dart';
+import '/widgets/dropdown_container.dart';
 import '/widgets/simple_dropdown_selector.dart';
 import 'providers/simple_dropdown_provider.dart';
 
-class DfSimpleDropdown<T> extends StatelessWidget {
-  /// Constructor for [DfSimpleDropdown].
+class DfDropdownWrapper<T> extends StatelessWidget {
+  /// Constructor for [DfDropdownWrapper].
   ///
   /// - [initData]: Initial list of data for the dropdown.
   /// - [selectedValue]: Currently selected dropdown value.
   /// - [labelText]: Text for the label of the dropdown.
-  /// - [hintText]: Placeholder text shown when no value is selected.
   /// - [onOptionSelected]: Callback function triggered when an option is selected.
   /// - [validator]: Optional validation function for dropdown selection.
   /// - [decoration]: Custom styling for the dropdown field.
@@ -23,23 +21,19 @@ class DfSimpleDropdown<T> extends StatelessWidget {
   /// - [arrowWidget]: Widget for the arrow icon displayed in the dropdown.
   /// - [dropdownType]: Default value is `DropdownType.expandable`, and it's used to switch between the expandable, and
   /// the overlay appearance
-  const DfSimpleDropdown({
+  const DfDropdownWrapper({
     super.key,
     this.initData = const [],
     this.selectedValue,
     this.labelText,
-    this.hintText,
     this.onOptionSelected,
     this.validator,
     this.decoration,
     this.selectorDecoration,
     this.arrowWidget,
-    this.dropdownType = DropdownType.expandable,
+    this.child,
+    this.closeOnTapOutside = true,
   });
-
-  ///Default value is `DropdownType.expandable`, and it's used to switch between the expandable, and
-  /// the overlay appearance
-  final DropdownType dropdownType;
 
   /// Initial list of dropdown options.
   final List<DropDownModel<T>> initData;
@@ -49,9 +43,6 @@ class DfSimpleDropdown<T> extends StatelessWidget {
 
   /// The label text for the dropdown field.
   final String? labelText;
-
-  /// Placeholder text displayed when no value is selected.
-  final String? hintText;
 
   /// Callback triggered when an option from the dropdown is selected.
   final Function(DropDownModel<T>)? onOptionSelected;
@@ -68,6 +59,10 @@ class DfSimpleDropdown<T> extends StatelessWidget {
   /// Widget displayed for the dropdown arrow icon.
   final Widget? arrowWidget;
 
+  final Widget? child;
+
+  final bool closeOnTapOutside;
+
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
@@ -83,9 +78,9 @@ class DfSimpleDropdown<T> extends StatelessWidget {
         arrowWidget: arrowWidget,
         selectorDecoration: selectorDecoration,
         decoration: decoration,
-        hintText: hintText,
         labelText: labelText,
-        dropdownType: dropdownType,
+        closeOnTapOutside: closeOnTapOutside,
+        child: child,
       ),
     );
   }
@@ -94,18 +89,18 @@ class DfSimpleDropdown<T> extends StatelessWidget {
 class _Dropdown<T> extends StatefulWidget {
   const _Dropdown({
     this.labelText,
-    this.hintText,
     required this.decoration,
     required this.selectorDecoration,
     required this.arrowWidget,
-    required this.dropdownType,
+    required this.child,
+    required this.closeOnTapOutside,
   });
   final SimpleSelectorDecoration? selectorDecoration;
   final DropdownDecoration? decoration;
   final String? labelText;
-  final String? hintText;
   final Widget? arrowWidget;
-  final DropdownType dropdownType;
+  final Widget? child;
+  final bool closeOnTapOutside;
 
   @override
   State<_Dropdown<T>> createState() => _DropdownState<T>();
@@ -116,68 +111,82 @@ class _DropdownState<T> extends State<_Dropdown<T>> {
 
   late SimpleDropdownProvider<T> provider;
 
+  bool tapOutside = false;
+
   @override
   void initState() {
     provider = Provider.of<SimpleDropdownProvider<T>>(context, listen: false);
     selectorWidget = Consumer<SimpleDropdownProvider<T>>(
-      builder: (_, provider, __) => SimpleDropdownSelector<T>(
-        selectorDecoration: widget.selectorDecoration,
-        selectedOption: provider.selectedValue,
-        dropdownData: provider.suggestionsExpanded ? provider.initData : [],
-        dropdownHeight: provider.dropdownHeight,
-        onSelectSuggestion: provider.onSelectSuggestion,
+      builder: (_, provider, __) => TapRegion(
+        onTapOutside: provider.suggestionsExpanded && widget.closeOnTapOutside
+            ? (_) {
+                tapOutside = true;
+                provider.closeSuggestions();
+                Future.delayed(const Duration(milliseconds: 200), () {
+                  tapOutside = false;
+                });
+              }
+            : null,
+        child: SimpleDropdownSelector<T>(
+          expanded: false,
+          selectedOption: provider.selectedValue,
+          selectorDecoration: widget.selectorDecoration,
+          dropdownData: provider.suggestionsExpanded ? provider.initData : [],
+          dropdownHeight: provider.dropdownHeight,
+          onSelectSuggestion: provider.onSelectSuggestion,
+        ),
       ),
     );
     super.initState();
 
-    if (widget.dropdownType == DropdownType.overlay) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        provider.updateSelectorPositionIfNeeded(
-          selectorWidget: ChangeNotifierProvider.value(
-            value: provider,
-            child: selectorWidget,
-          ),
-        );
-      });
-    }
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      provider.updateSelectorPositionIfNeeded(
+        expanded: false,
+        selectorWidget: ChangeNotifierProvider.value(
+          value: provider,
+          child: selectorWidget,
+        ),
+      );
+    });
   }
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        DropdownField<SimpleDropdownProvider<T>>(
-          key: provider.dropdownKey,
-          decoration: widget.decoration,
-          hintText: widget.hintText,
-          labelText: widget.labelText,
-          disableInput: true,
-          outlineBorderVisible: provider.suggestionsExpanded,
-          onTapInside: () => provider.toggleSuggestionsExpanded(
-            selectorWidget: widget.dropdownType == DropdownType.expandable
-                ? null
-                : ChangeNotifierProvider.value(
-                    value: provider,
-                    child: selectorWidget,
-                  ),
-          ),
-          suffixWidget: SizedBox(
-            height: 48,
-            child: widget.arrowWidget ??
-                Icon(
+    return DropdownContainer<SimpleDropdownProvider<T>>(
+      contentPadding: const EdgeInsets.all(0),
+      decoration: widget.decoration,
+      labelText: widget.labelText,
+      disableInput: true,
+      outlineBorderVisible: provider.suggestionsExpanded,
+      suffixTapEnabled: true,
+      suffixWidget: InkWell(
+        key: provider.dropdownKey,
+        onTap: () {
+          if (!tapOutside) {
+            provider.toggleSuggestionsExpanded(
+              expanded: false,
+              selectorWidget: ChangeNotifierProvider.value(
+                value: provider,
+                child: selectorWidget,
+              ),
+            );
+          }
+        },
+        child: SizedBox(
+          child: widget.arrowWidget ??
+              Padding(
+                padding: const EdgeInsets.only(
+                  right: 8,
+                ),
+                child: Icon(
                   context.watch<SimpleDropdownProvider<T>>().suggestionsExpanded
                       ? Icons.keyboard_arrow_up_outlined
                       : Icons.keyboard_arrow_down_outlined,
                 ),
-          ),
+              ),
         ),
-        if (widget.dropdownType == DropdownType.expandable) ...[
-          const SizedBox(
-            height: 8,
-          ),
-          selectorWidget,
-        ]
-      ],
+      ),
+      child: widget.child,
     );
   }
 }