+++ lib/widgets/simple_dropdown_selector.dart
@@ -9,13 +9,17 @@ class SimpleDropdownSelector<T> extends StatelessWidget {
     required this.dropdownData,
     required this.dropdownHeight,
     required this.onSelectSuggestion,
+    required this.selectedOption,
     this.selectorDecoration,
+    this.expanded = true,
   });
 
   final List<DropDownModel<T>> dropdownData;
   final double dropdownHeight;
   final SimpleSelectorDecoration? selectorDecoration;
   final Function(DropDownModel<T>) onSelectSuggestion;
+  final bool expanded;
+  final DropDownModel<T>? selectedOption;
 
   @override
   Widget build(BuildContext context) {
@@ -31,16 +35,19 @@ class SimpleDropdownSelector<T> extends StatelessWidget {
               selectorDecoration?.borderRadius ?? BorderRadius.circular(12),
           color: selectorDecoration?.selectorColor ?? Colors.white,
         ),
-        width: double.infinity,
+        width: expanded ? double.infinity : null,
         height: dropdownHeight,
         child: SingleChildScrollView(
           child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
             mainAxisSize: MainAxisSize.min,
             children: dropdownData
                 .map(
                   (suggestion) => _DropdownSuggestion(
+                    expanded: expanded,
                     selectorDecoration: selectorDecoration,
                     text: suggestion.text,
+                    selected: suggestion == selectedOption,
                     onTap: () {
                       onSelectSuggestion(suggestion);
                     },
@@ -65,17 +72,23 @@ class _DropdownSuggestion extends StatelessWidget {
     required this.text,
     required this.onTap,
     required this.selectorDecoration,
+    required this.expanded,
+    required this.selected,
   });
 
   final String text;
   final VoidCallback onTap;
   final SimpleSelectorDecoration? selectorDecoration;
+  final bool expanded;
+  final bool selected;
 
   @override
   Widget build(BuildContext context) {
     final textTheme = Theme.of(context).textTheme;
     return Material(
-      color: Colors.transparent,
+      color: selected
+          ? selectorDecoration?.selectedItemColor ?? Colors.transparent
+          : Colors.transparent,
       borderRadius:
           selectorDecoration?.borderRadius ?? BorderRadius.circular(12),
       child: InkWell(
@@ -90,11 +103,27 @@ class _DropdownSuggestion extends StatelessWidget {
                 selectorDecoration?.borderRadius ?? BorderRadius.circular(12),
             color: selectorDecoration?.itemColor ?? Colors.transparent,
           ),
-          width: double.infinity,
-          child: Text(
-            text,
-            style: selectorDecoration?.optionTextStyle ?? textTheme.labelMedium,
-            textAlign: TextAlign.start,
+          width: expanded
+              ? double.infinity
+              : selectorDecoration?.selectorWidth ?? 164,
+          child: Row(
+            children: [
+              Expanded(
+                child: Text(
+                  text,
+                  style: selectorDecoration?.optionTextStyle ??
+                      textTheme.labelMedium,
+                  textAlign: TextAlign.start,
+                  overflow: TextOverflow.ellipsis,
+                ),
+              ),
+              if (selectorDecoration?.selectedItemIcon != null && selected) ...[
+                const SizedBox(
+                  width: 4,
+                ),
+                selectorDecoration!.selectedItemIcon!,
+              ]
+            ],
           ),
         ),
       ),